Overview txt multilevel radar working (not exporting separate dds filescurrently) + minor

+++ JERC/Constants/Sizes.cs
@@ -6,7 +6,7 @@ namespace JERC.Constants
 {
     public static class Sizes
     {
-        public static readonly int MaxHammerGridSize = 32768;
+        public static readonly float MaxHammerGridSize = 32768.000000f;
         public static readonly int FinalOutputImageResolution = 1024;
         public static readonly int SizeReductionMultiplier = 1; // this should use scale from the overview txt
     }
